requirements.txt
alembic==1.16.5
bcrypt==4.3.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==2.3.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==2.3.7